requirements.txt
asgiref==3.3.1
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Markdown==3.3.3
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.5
sqlparse==0.4.1
